% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_replace}
\alias{find_replace}
\title{A Function to Edit Column Headers}
\usage{
find_replace(df, find, replace)
}
\description{
This function allows for finding and replacing patterns in column names.
}
\examples{
find_replace(df, find = "January 2018", replace = "2018-01")
}
\keyword{find,}
\keyword{rename}
\keyword{replace,}
