% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_import.R
\name{multi_import}
\alias{multi_import}
\title{A Nanosight Import Function from Multiple Folders}
\usage{
multi_import(dir, range, bin_width, NTA_version)
}
\arguments{
\item{dir}{The parent directory where folder and subsequent files are located, defaults to global working directory}

\item{range}{The range of values used during NTA data acquisition, defaults to 1000.}

\item{bin_width}{The bin width used during NTA data acquisition, defaults to 1.}

\item{NTA_version}{Version of NTA software (2.3, 3.1, 3.2).}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows for the import and combining of multiple csv files within multiple folders into a single data frame.
}
\examples{
multi_import(dir = "parent_folder", bin_width = 1, range = 1000, NTA_version = 3.2)
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
