% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoANOVA.R
\name{nanoANOVA}
\alias{nanoANOVA}
\title{A nested ANOVA Function}
\usage{
nanoANOVA(df, ..., name = "Param", value, variables)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Columns to group by.}

\item{value}{The numerical value that you want to summarize.}

\item{variable}{The variable you wish to compare between.}
}
\value{
A nested data frame grouped by \code{...} and columns for Shapiro Test,
ANOVA and Tukey HSD post hoc test.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanoANOVA(starwars, species, param_var = mass)
nanoANOVA(starwars, gender, param_var = height)
}
\keyword{ANOVA,}
\keyword{nested}
\keyword{statistics,}
