% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocombine.R
\name{nanocombine}
\alias{nanocombine}
\title{A Nanosight  Import Function}
\usage{
nanocombine(dir = "", auto_name = FALSE, custom_name = NULL)
}
\arguments{
\item{dir}{The directory where files are located, defaults to global working directory}

\item{auto_name}{Extracts sample and dilution data and appends to column header, defaults to FALSE.}

\item{custom_name}{Append custom name to column header, defaults to NULL.}
}
\value{
Dataframe of NTA data.
}
\description{
This function allows you import and combine multiple data files into one data frame.
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
