% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocount.R
\name{nanocount}
\alias{nanocount}
\title{A Sum Function}
\usage{
nanocount(df, ..., name = "Particle", param_var, na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Columns to group by.}

\item{name}{Prefix name to sum function.}

\item{param_var}{The numerical value that you want to summarize.}
}
\value{
Sum of particle count \code{parm_var} grouped by \code{...}.
Output of tidy dataframe of  sums grouped by \code{...} labeled according to 'name' argument.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanocount(starwars, species,name = "mass", param_var = mass)
nanocount(starwars, gender,name = "height", param_var = height)
}
\keyword{count,}
\keyword{sum}
