% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoimport.R
\name{nanoimport}
\alias{nanoimport}
\title{A Nanosight Data Import Function}
\usage{
nanoimport(file, auto_name = FALSE, custom_name = NULL)
}
\arguments{
\item{file}{A .csv file from a Malvern Nanosight machine.}

\item{auto_name}{Extracts sample and dilution data and appends to column header, defaults to FALSE.}

\item{custom_name}{Append custom name to column header, defaults to NULL.}
}
\value{
Dataframe of NTA data.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
