% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanolyze.R
\name{nanolyze}
\alias{nanolyze}
\title{A Summarization Function}
\usage{
nanolyze(df, ..., name = "Param", param_var, na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Columns to group by.}

\item{name}{Prefix name to summary statistics}

\item{param_var}{The numerical value that you want to summarize.}

\item{na.rm}{Argument to ignore 'NA' values, defaults to TRUE.}
}
\value{
Summarized values \code{parm_var} grouped by \code{...}.
Output of tidy dataframe of grouped by \code{...} variables, N, mean,
standard deviation and standard error labeled according to 'name' argument.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanolyze(mtcars, cyl ,name = "hp", param_var = hp)
nanolyze(iris,Species, name = "Sepal_Width", param_var = Sepal.Width)
}
\keyword{aggregate,}
\keyword{statistics}
\keyword{summarize,}
\keyword{summary}
