% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanonest.R
\name{nanonest}
\alias{nanonest}
\title{A Function for Nesting Data}
\usage{
nanonest(df, ...)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Columns to group by.}
}
\value{
A nested data frame grouped by \code{...} , suitable for using mutate and map to
perform statistical tests.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanonest(starwars, species)
nanonest(starwars, gender)
}
\keyword{nested}
\keyword{nesting,}
\keyword{statistics,}
