% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanosave.R
\name{nanosave}
\alias{nanosave}
\title{A Function to Save Tidy Data Frames}
\usage{
nanosave(df, name = "tidyNano")
}
\arguments{
\item{df}{file data frame.}

\item{name}{The name of the Rds file, defaults to "tidyNano.Rds".}
}
\description{
This function exports data frames into Rds files.
}
\examples{
nanosave(iris, name = "iris")
}
\keyword{export}
\keyword{save,}
