% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoShapiro.R
\name{nanoShapiro}
\alias{nanoShapiro}
\title{A nested Normality Test}
\usage{
nanoShapiro(df, ..., value, variables)
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Columns to group by.}

\item{value}{The numerical value that you want to summarize.}
}
\value{
A nested data frame grouped by \code{...} and columns for Shapiro Test.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanoShapiro(starwars, species, param_var = mass)
nanoShapiro(starwars, gender, param_var = height)
}
\keyword{nested}
\keyword{normality,}
\keyword{parametric}
\keyword{statistics,}
\keyword{test,}
