% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocombine.R
\name{nanocombine}
\alias{nanocombine}
\title{A Nanosight  Import Function}
\usage{
nanocombine(dir = "", bin_width = 1, range = 1000, NTA_version)
}
\arguments{
\item{dir}{The directory where files are located, defaults to global working directory}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows you import and combine multiple data files into one data frame.
}
\examples{
nanocombine(dir = "file_folder", bin_width = 1, range = 1000, NTA_version = 3.2)
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
