% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocombine.R
\name{nanocombine}
\alias{nanocombine}
\title{A Nanosight  Import Function}
\usage{
nanocombine(dir = "", range = 1000, bin_width = 1, nm_start = 0.5,
  auto_name = FALSE, custom_name = NULL)
}
\arguments{
\item{dir}{The directory where files are located, defaults to global working directory}

\item{range}{The range of values used during NTA data acquisition, defaults to 1000.}

\item{bin_width}{The bin width used during NTA data acquisition, defaults to 1.}

\item{nm_start}{Starting measurement value (nm), defaults to 0.5nm}

\item{auto_name}{Extracts sample and dilution data and appends to column header, defaults to FALSE.}

\item{custom_name}{Append custom name to column header, defaults to NULL.}
}
\value{
Dataframe of NTA data.
}
\description{
This function allows you import and combine multiple data files into one data frame.
}
\examples{
nanocombine(dir = "file_folder")
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
