% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanocombine.R
\name{nanocombine}
\alias{nanocombine}
\title{A Nanosight  Import Function}
\usage{
nanocombine(dir = "", bin_width = 1, range = 1000, NTA_version,
  extra_param = NULL)
}
\arguments{
\item{dir}{The directory where files are located, defaults to global working directory}

\item{bin_width}{The bin width used during NTA data acquisition, defaults to 1.}

\item{range}{The range of values used during NTA data acquisition, defaults to 1000.}

\item{NTA_version}{Version of NTA software (2.3, 3.1, 3.2).}

\item{extra_param}{Presence of additional data beside counts (Surface Area and Volume weighting)}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows you import and combine multiple data files into one data frame.
}
\examples{
nanocombine(dir = "file_folder", bin_width = 1, range = 1000, NTA_version = 3.2)
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
