% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoimport.R
\name{nanoimport}
\alias{nanoimport}
\title{A Nanosight Data Import Function}
\usage{
nanoimport(file, bin_width = 1, range = 1000, NTA_version,
  extra_param = NULL)
}
\arguments{
\item{file}{A .csv file from a Malvern Nanosight machine.}

\item{bin_width}{The bin width used during NTA data acquisition, defaults to 1.}

\item{range}{The range of values used during NTA data acquisition, defaults to 1000.}

\item{NTA_version}{Version of NTA software (2.3,3.1,3.2).}

\item{extra_param}{Presence of additional data beside counts (Surface Area and Volume weighting)}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanoimport(file = "nanofile.csv", bin_width = 1,, range = 1000, NTA_version = 3.2)
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
