% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoimport.R
\name{nanoimport}
\alias{nanoimport}
\title{A Nanosight Data Import Function}
\usage{
nanoimport(file, bin_width = 1, NTA_version)
}
\arguments{
\item{file}{A .csv file from a Malvern Nanosight machine.}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanoimport(file = "nanofile.csv", bin_width = 1, NTA_version = 3.2)
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
