% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoimport.R
\name{nanoimport}
\alias{nanoimport}
\title{A Nanosight Data Import Function}
\usage{
nanoimport(file, range = 1000, bin_width = 1, nm_start = 0.5,
  auto_name = FALSE, custom_name = NULL)
}
\arguments{
\item{file}{A .csv file from a Malvern Nanosight machine.}

\item{range}{The range of values (nm) used during NTA data acquisition, defaults to 1000.}

\item{bin_width}{The bin width used during NTA data acquisition, defaults to 1.}

\item{nm_start}{Starting measurement value (nm), defaults to 0.5nm}

\item{auto_name}{Extracts sample and dilution data and appends to column header, defaults to FALSE.}

\item{custom_name}{Append custom name to column header, defaults to NULL.}
}
\value{
Dataframe of NTA data.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\keyword{extract}
\keyword{import,}
\keyword{load,}
