% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoimport.R
\name{nanoimport}
\alias{nanoimport}
\title{A Nanosight Data Import Function}
\usage{
nanoimport(file)
}
\arguments{
\item{csv}{A .csv file.}
}
\value{
Dataframe of nanosight values.
}
\description{
This function allows you aggregate or summarize your data by groups.
}
\examples{
nanoimport("nanofile.csv")
}
\keyword{aggregate,}
\keyword{statistics}
\keyword{summarize,}
\keyword{summary}
