% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotidy.R
\name{nanotidy}
\alias{nanotidy}
\title{A Tidy Function}
\usage{
nanotidy(df, sep_var)
}
\arguments{
\item{df}{A dataframe.}

\item{sep_var}{Numerical range of columns to tidy, specify "2:n" where n is the last column}

\item{col}{Specify the range of columns to tidy (x:z).}
}
\value{
Gathers values from 'wide' to 'long' format. Returns a dataframe with a
Sample and Count column.
}
\description{
This function allows you tidy Nanosight data.  It will make the data go from
'wide' to 'long'.
}
\examples{
nanotidy(data, col = 2:12, sep_var = c("Sample","Dilution","Treatment","Condition"))
}
\keyword{aggregate,}
\keyword{clean,}
\keyword{tidy}
