% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotidy.R
\name{nanotidy}
\alias{nanotidy}
\title{A Tidy Function}
\usage{
nanotidy(df, ..., col_names, sep = "_")
}
\arguments{
\item{df}{A dataframe.}

\item{...}{Specify the range of columns to tidy (x:z).}
}
\value{
Gathers values from 'wide' to 'long' format. Returns a dataframe with a
Sample and Count column.
}
\description{
This function allows you tidy Nanosight data.  It will make the data go from
'wide' to 'long'.
}
\examples{
nanotidy(data, 2:12)
}
\keyword{aggregate,}
\keyword{statistics}
\keyword{summarize,}
\keyword{summary}
